third_party/WebKit/LayoutTests/svg/text/selection-dragging-outside-2.html

<!DOCTYPE html>
<style>
text {
  font: 10px Arial;
}
</style>
<svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" width="400" height="400" viewBox="0 0 500 500" style="border: 1px solid black;">
  <g>
    <text id="line1" x="10" y="30">Hello World. Hello, SVG.</text>
  </g>
  <g transform="translate(120, 30)">
    <text id="line2" x="50" y="50">Hello World. Hello, SVG.</text>
  </g>
  <g transform="translate(240, 30)">
    <text id="line3" x="100" y="100">Hello World. Hello, SVG.</text>
  </g>
</svg>
<script src="../../resources/js-test.js"></script>
<script src="resources/SelectionTestCase.js"></script>
<script type="text/javascript">
function getEndPosition(id, offset, gap) {
  var element = document.getElementById(id);
  var endPos = element.getEndPositionOfChar(offset);
  endPos.x += gap.x;
  endPos.y += gap.y;
  var absEndPos = toAbsoluteCoordinates(endPos, element);
  return absEndPos;
}

var absEndPos = getEndPosition('line3', 10, { x : 0, y : 30});
selectTextFromCharToPoint({ id: 'line1', offset: 7 }, absEndPos, { startElementId: "line1", start: "7", endElementId: "line3", end: "11" });

absEndPos = getEndPosition('line2', 10, { x : 0, y : 30});
selectTextFromCharToPoint({ id: 'line1', offset: 7 }, absEndPos, { startElementId: "line1", start: "7",  endElementId: "line2", end: "11" });

absEndPos = getEndPosition('line3', 10, { x : 0, y : 30});
selectTextFromCharToPoint({ id: 'line2', offset: 7 }, absEndPos, { startElementId: "line2", start: "7",  endElementId: "line3", end: "11" });

absEndPos = getEndPosition('line1', 8, { x : 0, y : -30});
selectTextFromCharToPoint({ id: 'line3', offset: 7 }, absEndPos, { startElementId: "line3", start: "7",  endElementId: "line1", end: "9" });

absEndPos = getEndPosition('line1', 8, { x : 0, y : -30});
selectTextFromCharToPoint({ id: 'line2', offset: 7 }, absEndPos, { startElementId: "line2", start: "7",  endElementId: "line1", end: "9" });

absEndPos = getEndPosition('line2', 8, { x : 0, y : -30});
selectTextFromCharToPoint({ id: 'line3', offset: 7 }, absEndPos, { startElementId: "line3", start: "7",  endElementId: "line2", end: "9" });
</script>
